+++ src/services/EventListener.php
@@ -2,20 +2,24 @@
 
 namespace kbergha\revisjonsspor\services;
 
+use craft\services\Users;
 use kbergha\revisjonsspor\Revisjonsspor;
 use yii\base\Component;
 use yii\base\Event;
 use yii\web\User;
 use yii\web\UserEvent;
 use craft\events\ElementEvent;
+use craft\events\UserGroupsAssignEvent;
 use craft\services\Elements;
-use craft\base\ElementInterface;
 use craft\base\Element;
 use craft\helpers\ElementHelper;
 use Craft;
 
 class EventListener extends Component
 {
+    /**
+     * Add listeners for various events.
+     */
     public function addEventListeners() {
 
         /*
@@ -37,6 +41,24 @@ class EventListener extends Component
             }
         );
 
+
+        Event::on(
+            Users::class,
+            Users::EVENT_AFTER_ACTIVATE_USER,
+            function($event) {
+                //$this->onLogout(Users::EVENT_AFTER_ACTIVATE_USER);
+            }
+        );
+
+        Event::on(
+            Users::class,
+            Users::EVENT_AFTER_ASSIGN_USER_TO_GROUPS,
+            function(UserGroupsAssignEvent $event) {
+                $this->onAssignedUserGroups(Users::EVENT_AFTER_ASSIGN_USER_TO_GROUPS, $event);
+            }
+        );
+
+        // @todo Users::EVENT_AFTER_ACTIVATE_USER ??
 
         /*
          * Elements events (Assets, Users, Entries, Categories +++)
@@ -62,9 +84,33 @@ class EventListener extends Component
 
         /*
          * Some other events
-         * */
+         */
     }
 
+    /**
+     * Whan a user is added or removed from user groups.
+     *
+     * @param $eventName
+     * @param UserGroupsAssignEvent $event
+     */
+    public function onAssignedUserGroups($eventName, UserGroupsAssignEvent $event)
+    {
+
+        $userGroupString = 'was assigned user groups with id '.implode(', ', $event->groupIds);
+        if (count($event->groupIds) === 0) {
+            $userGroupString = 'was removed from all user groups';
+        }
+
+        $properties = $this->getDefaultProperties($eventName);
+        $message = 'User with id '.$event->userId .' '.$userGroupString.' by '.$properties['userName'].' (id: '.$properties['userId'].')';
+        $this->log($message);
+    }
+
+    /**
+     * When a user logs in.
+     *
+     * @param $eventName
+     */
     public function onLogin($eventName)
     {
         $properties = $this->getDefaultProperties($eventName);
@@ -72,6 +118,11 @@ class EventListener extends Component
         $this->log($message, $properties);
     }
 
+    /**
+     * When a user logs out.
+     *
+     * @param $eventName
+     */
     public function onLogout($eventName)
     {
         $properties = $this->getDefaultProperties($eventName);
@@ -79,6 +130,14 @@ class EventListener extends Component
         $this->log($message, $properties);
     }
 
+    /**
+     * Do something useful when an element is saved
+     *
+     * @param $eventName
+     * @param Element $element
+     * @param bool $isNew
+     * @return bool
+     */
     public function onSaveElement($eventName, Element $element, $isNew = false)
     {
         // Don't log drafts/revisions and propagating/resaving elements
@@ -98,10 +157,23 @@ class EventListener extends Component
             $status = 'created';
         }
 
-        $this->log('Element of type "'.$element->displayName(). '" (id: '.$properties['elementId'].') was '.$status.' by user '.$properties['userName'].' (id: '.$properties['userId'].')', $properties);
+        $userInfo = '';
+        if ($element instanceof \craft\elements\User) {
+            /* @var $element \craft\elements\User */
+            $userInfo = 'username: '.$element->username.', email: '. $element->email.', ';
+        }
+
+        $this->log('Element of type "'.$element->displayName(). '" ('.$userInfo.'id: '.$properties['elementId'].') was '.$status.' by user '.$properties['userName'].' (id: '.$properties['userId'].')');
     }
 
 
+    /**
+     * Do something useful when an element is deleted.
+     *
+     * @param $eventName
+     * @param Element $element
+     * @return bool
+     */
     public function onDeleteElement($eventName, Element $element)
     {
         // Don't log drafts/revisions
@@ -115,11 +187,19 @@ class EventListener extends Component
         $properties['elementUid'] = $element->getSourceUid();
         $properties['class'] = get_class($element);
 
-        $this->log('Element of type "'.$element->displayName(). '" (id: '.$properties['elementId'].') was deleted by user '.$properties['userName'].' (id: '.$properties['userId'].')', $properties);
+        $userInfo = '';
+        if ($element instanceof \craft\elements\User) {
+            /* @var $element \craft\elements\User */
+            $userInfo = 'username: '.$element->username.', email: '. $element->email.', ';
+        }
+
+        $this->log('Element of type "'.$element->displayName(). '" ('.$userInfo.'id: '.$properties['elementId'].') was deleted by user '.$properties['userName'].' (id: '.$properties['userId'].')');
     }
 
 
     /**
+     * Get an array of various useful properties that can be used for various useful things.
+     *
      * @return array
      */
     protected function getDefaultProperties($eventName) : array
@@ -149,14 +229,23 @@ class EventListener extends Component
         return $properties;
     }
 
-    protected function log($message, $properties)
+    /**
+     * Log to whatever Craft is configured to log to.
+     *
+     * @param $message
+     * @param null $properties
+     */
+    protected function log($message, $properties = null)
     {
-        $settings = Revisjonsspor::$plugin->getSettings();
-        // @todo: make properties logging configurable?
-        // @todo: try-catch
+        try {
+            $settings = Revisjonsspor::$plugin->getSettings();
+            if (is_array($properties) && count($properties) > 0) {
+                $message .= ' - props: '.\json_encode($properties);
+            }
 
-        $message .= ' - props: '.\json_encode($properties);
-
-        Craft::getLogger()->log($message, $settings['level'], $settings['category']);
+            Craft::getLogger()->log($message, $settings['level'], $settings['category']);
+        } catch (\Exception $e) {
+            Craft::getLogger()->log('Could not log to audit log, exception error message: '.$e->getMessage(), Craft::getLogger()::LEVEL_ERROR);
+        }
     }
 }