+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.0-alpha2 - 2020-04-20
+### Added
+- Adds logging of user group assignments.
+- Added email and username of edited or saved user to log.
+- Error handling in the log function
+- Some doc blocks / code comments.
+
 ## 1.0.0-alpha - 2020-04-01
 ### Added
 - Initial release